Preserves permissions when renaming files.
Previously the renaming of files after translation was reimplemented as "rm" followed by "cp" to workaround a bug in ther virtual machine used by Docker toolbox on windows 10 home. But in unix, since the docker process runs as root, this would produce a copy of the file associated with user root.  Now we use "cp -p" to preserve permissions on copy.